chembl/requirements.txt
numpy==1.26.1
pandas==2.1.1
psycopg2-binary==2.9.9
PubChemPy==1.0.4
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
